Hit a wall -- Need to generalize and solve ptr to nil reference vs nil ptr to reference.

+++ eco/keyValueTree.go
@@ -11,8 +11,8 @@ import (
 
 type KeyValueChildMap map[KeyString]*KeyValueTree
 
-func (m KeyValueChildMap) MaxIndex() uint64 {
-	var maxIndex uint64 = 0
+func (m KeyValueChildMap) MaxIndex() int {
+	var maxIndex int = 0
 	for key := range m {
 		keyString := KeyString(key)
 		index, is := keyString.Index()
